+++ onnxruntime/core/providers/dml/DmlExecutionProvider/src/Operators/bfx/grid_sample.h
@@ -94,7 +94,7 @@ public:
                 ML_CHECK_VALID_ARGUMENT(rank == 4, "Offset shape must be 4D.");
                 auto dims = shapeInfo.GetInputTensorShape(1);
                 assert(dims.size() == rank);
-                assert(n == dims[0]);
+                assert(n == static_cast<int32_t>(dims[0]));
                 assert(2 == dims[1]);
                 out_h = static_cast<int32_t>(dims[2]);
                 out_w = static_cast<int32_t>(dims[3]);
